refactor code

+++ src/main/java/com/example/demo/dto/SearchProductRequest.java
@@ -12,8 +12,8 @@ import lombok.Setter;
 public class SearchProductRequest {
     private String name;
     private Integer age;
-    private Double minPrice;
-    private Double maxPrice;
+    private Float minPrice;
+    private Float maxPrice;
     private Long productTypeId;
     private Integer status;
